% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrech.R
\name{constrech}
\alias{constrech}
\title{Construit un echantillon}
\usage{
constrech(faisabl, data, constr, methode)
}
\arguments{
\item{faisabl}{resultats de la fonction faisable()}

\item{data}{un jeu de donnée}

\item{constr}{un tableau de quota de type df avec comme variable "variables", "modalitesé" et "objectifs"}

\item{methode}{True pour l'approche probabiliste et False pour l'approche par optimisation}
}
\value{
un echantillon
}
\description{
construit un echantillon selon la methode des quotas avec la methode à deux niveaux
}
\examples{
#constr=data.frame(variables=c("Species","Species","Species"),
#modalites=c("setosa","versicolor","virginica"),objectifs=c(25,25,40))
#faisabl=faisable(iris,constr)
#iris$risque<-sample(seq(0,1,0.01),nrow(iris),TRUE)
#iris$tauxrep<-sample(seq(0,1,0.01),nrow(iris),TRUE)
#constrech(faisabl,iris,constr,TRUE)
#constrech(faisabl,iris,constr,FALSE)
}
