% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrsamp.R
\name{constrsamp}
\alias{constrsamp}
\title{Construit un echantillon}
\usage{
constrsamp(data, constr, foyer, nb)
}
\arguments{
\item{data}{un jeu de donnée}

\item{constr}{un tableau de quota de type df avec comme variable "variables", "modalitesé" et "objectifs"}

\item{foyer}{varaible qui sert d'identifiant foyer}

\item{nb}{taille de l'echantillon}
}
\value{
un echantillon
}
\description{
construit un echantillon selon la methode des quotas avec la methode pas-à-pas
}
\examples{
#Pas d'exemple
}
