% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremum.R
\name{extremum}
\alias{extremum}
\title{extremum}
\usage{
extremum(sysineq, direction, numeros)
}
\arguments{
\item{sysineq}{systeme d'inequation auquelle est rattache une variable}

\item{direction}{"min" pour borne inf ou "max" pour borne sup}

\item{numeros}{numeros de la variable}
}
\value{
la borne sup ou inf de la variable
}
\description{
Permet de determiner la borne sup et inf de d'une variable
}
\examples{
#no exemple
}
