% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faisable.R
\name{faisable}
\alias{faisable}
\title{Faisable}
\usage{
faisable(data, constr, taillech = NA)
}
\arguments{
\item{data}{un jeu de donnée}

\item{constr}{un tableau de quota de type df avec comme variable "variables", "modalitesé" et "objectifs"}

\item{taillech}{taille de l'échantillon}
}
\value{
une liste avec la faisabilité, les résultas des systemes linéaires associés
}
\description{
determine la fasabilite d'un tableau de quota par rapport à un jeu de donnees
}
\examples{
#constr=data.frame(variables=c("Species","Species","Species"),
#modalites=c("setosa","versicolor","virginica"),objectifs=c(98,123,100))
#faisable(iris,constr)
}
