% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{fusion}
\alias{fusion}
\title{fusion}
\usage{
fusion(l)
}
\arguments{
\item{l}{une liste de matrice ou df}
}
\value{
un df ou matrice résultant de la fusion de ceux contenus dans l
}
\description{
Une fonction équivalente à rbind mais sur une liste de fonction ou de matrice.
}
\examples{
#fusion(list(matrix(1,5,5),matrix(0,5,5)))

}
