% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion2.R
\name{fusion2}
\alias{fusion2}
\title{fusion2}
\usage{
fusion2(l)
}
\arguments{
\item{l}{liste de df ou matrice}
}
\value{
la fusion en colone de la liste
}
\description{
fusionne les matrices ou df d'une liste à la manière de cbind
}
\examples{
fusion2(list(matrix(0,5,5),matrix(1,5,5)))
}
