% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianElimination.R
\name{gaussianElimination}
\alias{gaussianElimination}
\title{gaussianElimination
pivot de gauss}
\usage{
gaussianElimination(A, B, tol = sqrt(.Machine$double.eps),
  verbose = FALSE, latex = FALSE, fractions = FALSE)
}
\arguments{
\item{A}{matrice}

\item{B}{matrice}

\item{tol}{inutile}

\item{verbose}{inutile}

\item{latex}{inutile}

\item{fractions}{inutile}
}
\value{
matrice
}
\description{
gaussianElimination
pivot de gauss
}
\examples{
#no
}
