% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identique.R
\name{identique}
\alias{identique}
\title{identique}
\usage{
identique(a)
}
\arguments{
\item{a}{vecteur quelconque}
}
\value{
une liste avec les indice de chaque modalité
}
\description{
Renvoie une liste de meme taille que unique(a) avec les indices
}
\examples{
identique(c(1,2,3,1,3,2))
}
