% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insererlignes.R
\name{insererlignes}
\alias{insererlignes}
\title{Insererligne}
\usage{
insererlignes(data, NewRow, RowNum = NULL)
}
\arguments{
\item{data}{un jeu de donnees}

\item{NewRow}{un vecteur de m}

\item{RowNum}{numéro de la ligne}
}
\value{
data avec les lignes inserer au bon endroit
}
\description{
Insererligne
}
\examples{
#No exemple
}
