% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsyslin.R
\name{matsyslin}
\alias{matsyslin}
\title{matsyslin}
\usage{
matsyslin(data, constr)
}
\arguments{
\item{data}{un jeu de donnée}

\item{constr}{un tableau de quota de type df avec comme variable "variables", "modalitesé" et "objectifs"}
}
\value{
la matrice du systeme linéaire ass
}
\description{
matsyslin
}
\examples{
constr=data.frame(variables=c("Species","Species","Species"),
modalites=c("setosa","versicolor","virginica"),objectifs=c(98,123,100))
matsyslin(iris,constr)
}
