% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodeopt.R
\name{methodeopt}
\alias{methodeopt}
\title{methode par optimisation}
\usage{
methodeopt(foncoptm, faisabl)
}
\arguments{
\item{foncoptm}{vecteur de coefficient de la fonction a optimiser}

\item{faisabl}{resultat de la fonction faisable()}
}
\value{
un vecteur qui permet de determiner la configuration de l'echantillon
}
\description{
methode par optimisation
}
\examples{
#pas d'exemple
}
