% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodeprob.R
\name{methodeprob}
\alias{methodeprob}
\title{methode probabiliste}
\usage{
methodeprob(foncoptm, sysineq, R)
}
\arguments{
\item{foncoptm}{coefficient de la fonction a optimiser}

\item{sysineq}{systeme d'inequation lineaire}

\item{R}{vecteur de parametre}
}
\value{
returne un vecteur de parametre avec des valeurs
}
\description{
configure l'echantillion avec une approche probabiliste
}
\examples{
#no exemple
}
