% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication.R
\name{multiplication}
\alias{multiplication}
\title{multiplication}
\usage{
multiplication(m, l)
}
\arguments{
\item{m}{veceur de nombre}

\item{l}{vecteur de nombre}
}
\value{
un vecteur
}
\description{
effectue une multiplication
}
\examples{
#NOt exemple
}
