% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation_ech.R
\name{optimisation_ech}
\alias{optimisation_ech}
\title{Optimisation de la construction des echantillons avec la méthode à deux niveaux}
\usage{
optimisation_ech(l, methode)
}
\arguments{
\item{l}{resultat de la fonction optimisation_f()}

\item{methode}{TRUE pour l'approche probabiliste et FALSE pour l'approche par optimisation}
}
\value{
un echantillon
}
\description{
Optimisation de la construction des echantillons avec la méthode à deux niveaux
}
\examples{
iris$couleur=sample(c("rose","blanc"),150,replace = TRUE)
iris$risque<-sample(seq(0,1,0.01),nrow(iris),TRUE)
iris$tauxrep<-sample(seq(0,1,0.01),nrow(iris),TRUE)
constr=data.frame(variables=c("Species","Species","Species","couleur","couleur"),
modalites=c("setosa","versicolor","virginica","rose","blanc"),objectifs=c(25,25,25,35,40))
l=optimisation_f("couleur",iris,constr,75)
optimisation_ech(l,TRUE)
}
