% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation_f.R
\name{optimisation_f}
\alias{optimisation_f}
\title{Optimisation de faisable}
\usage{
optimisation_f(varselect, data, constr, taille)
}
\arguments{
\item{varselect}{variable utilise pour diviser la tableau de quota}

\item{data}{base de sondage}

\item{constr}{tableau de quota}

\item{taille}{taille de l'echantillon}
}
\value{
une liste avec la faisabilite, la liste des bases de sondages, la liste des sous-tableau
}
\description{
Cette fonction permet de determiner la faisabilité plus rapidement que la fonction faisable
}
\examples{
iris$couleur=sample(c("rose","blanc"),150,replace = TRUE)
constr=data.frame(variables=c("Species","Species","Species","couleur","couleur"),
modalites=c("setosa","versicolor","virginica","rose","blanc"),objectifs=c(25,25,25,35,40))
optimisation_f("couleur",iris,constr,75)
}
