% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba.R
\name{proba}
\alias{proba}
\title{proba}
\usage{
proba(poid, inf, sup)
}
\arguments{
\item{poid}{le poid de la variable dans la fonction à optimiser}

\item{inf}{borne inf de la variable}

\item{sup}{born sup de la variable}
}
\value{
une loi de probabilite
}
\description{
proba
}
\examples{
proba(0.9,0,100)
}
