% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduction}
\alias{reduction}
\title{Reduction}
\usage{
reduction(coeff, constantes, sup = NA, inf = NA)
}
\arguments{
\item{coeff}{matrice de coefficients du système}

\item{constantes}{vecteur de constantes du systeme}

\item{sup}{vecteur de bornes sup}

\item{inf}{vecteur de bornes inf}
}
\value{
retourne une liste avec la compatibilité du systeme et les formes reduit du systeme
}
\description{
Reduit un systeme d'inequations linéaires et determine sa compatibilite
}
\examples{
#coeff=matrix(c(1,1,1,0,0,1),byrow=TRUE,ncol=2,nrow=3)
#sup=c(50,75,25)
#inf=c(12,10,15)
#constantes=c(12,2,3)
#reduction(coeff,constantes,sup,inf)
}
