% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduire.R
\name{reduire}
\alias{reduire}
\title{reduire}
\usage{
reduire(numeros, valeur, sysineq)
}
\arguments{
\item{numeros}{numeros de la variable a qui on vient d'attribuer une valeur}

\item{valeur}{valeur qu'on attribue à la variable}

\item{sysineq}{le systeme d'inequation associe}
}
\value{
un systeme d'inequation reduit
}
\description{
permet de reduire le systeme d'inequation aprs l attribution d'une valeur à une variable
}
\examples{
#no exemple
}
