% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvequ.R
\name{solvequ}
\alias{solvequ}
\title{Solvequ}
\usage{
solvequ(matsys, b)
}
\arguments{
\item{matsys}{matrice de coefficient du systeme lineaire}

\item{b}{vecteur d'egalies du systeme lineaire}
}
\value{
une liste qui indique si le systeme est compatible ainsi que les solutions
}
\description{
function qui facilite la résoltion de systeme lineaire
}
\examples{
matsys=matrix(c(1,1,5,1,0,1),byrow=TRUE,ncol=3,nrow=2)
b=c(25,5)
solvequ(matsys,b)
}
