% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerosbis.R
\name{zerosbis}
\alias{zerosbis}
\title{Zerosbis}
\usage{
zerosbis(n, m = n)
}
\arguments{
\item{n}{nombre}

\item{m}{nombre}
}
\value{
matrice de zéros
}
\description{
crée une matrice de zéros
}
\examples{
zerosbis(5.5,7.8)
}
